+++ novacore/src/main/java/com/starlwr/bot/core/event/live/common/RandomGiftEvent.java
@@ -5,7 +5,7 @@ import com.starlwr.bot.core.event.live.base.StarBotLiveGiftEvent;
 import com.starlwr.bot.core.model.GiftInfo;
 import com.starlwr.bot.core.model.LiveStreamerInfo;
 import com.starlwr.bot.core.model.UserInfo;
-import com.starlwr.bot.core.util.MathUtil;
+import com.starlwr.bot.core.lang.MathUtil;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;