+++ novacore/src/main/java/com/starlwr/bot/core/lang/StringUtil.java
@@ -1,4 +1,4 @@
-package com.starlwr.bot.core.util;
+package com.starlwr.bot.core.lang;
 
 /**
  * 字符串工具类