+++ starbot-bilibili/src/main/java/com/starlwr/bot/bilibili/handler/PushHandlerSupport.java
@@ -7,7 +7,7 @@ import com.starlwr.bot.core.model.Message;
 import com.starlwr.bot.core.model.PushTarget;
 import com.starlwr.bot.core.sender.AtMode;
 import com.starlwr.bot.core.sender.StarBotMessageSender;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;