+++ starbot-core/src/main/java/com/starlwr/bot/core/listener/StarBotHandlerListener.java
@@ -9,7 +9,7 @@ import com.starlwr.bot.core.model.PushUser;
 import com.starlwr.bot.core.sender.PushGate;
 import com.starlwr.bot.core.timeline.TimelineEvent;
 import com.starlwr.bot.core.timeline.TimelineWriter;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.EventListener;