+++ starbot-core/src/main/java/com/starlwr/bot/core/sender/PushGate.java
@@ -2,7 +2,7 @@ package com.starlwr.bot.core.sender;
 
 import com.starlwr.bot.core.config.StarBotCoreProperties;
 import com.starlwr.bot.core.timeline.TimelineEventType;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;