+++ starbot-core/src/main/java/com/starlwr/bot/core/service/StarBotMailService.java
@@ -1,7 +1,7 @@
 package com.starlwr.bot.core.service;
 
 import com.starlwr.bot.core.config.StarBotCoreProperties;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import jakarta.mail.internet.MimeMessage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.ObjectProvider;