+++ starbot-onebot-adapter/src/main/java/com/starlwr/bot/adapter/onebot/controller/OneBotController.java
@@ -10,7 +10,7 @@ import com.starlwr.bot.core.enums.PushTargetType;
 import com.starlwr.bot.core.model.Sender;
 import com.starlwr.bot.core.plugin.StarBotComponent;
 import com.starlwr.bot.core.service.StarBotSenderService;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;