+++ starbot-report/src/main/java/com/starlwr/bot/report/painter/BilibiliLiveReportPainter.java
@@ -20,7 +20,7 @@ import com.starlwr.bot.core.model.UserScore;
 import com.starlwr.bot.core.plugin.StarBotComponent;
 import com.starlwr.bot.core.service.LiveDataService;
 import com.starlwr.bot.core.service.LiveRoomInfoHistory;
-import com.starlwr.bot.core.util.StringUtil;
+import com.starlwr.bot.core.lang.StringUtil;
 import com.starlwr.bot.report.factory.StarBotCommonPainterFactory;
 import com.starlwr.bot.report.util.FontUtil;
 import com.starlwr.bot.report.util.ImageUtil;